+++ Products/PloneLDAP/plugins/base.py
@@ -10,8 +10,7 @@ from Products.PluggableAuthService.interfaces.plugins import \
     IAuthenticationPlugin, IRolesPlugin, \
     ICredentialsResetPlugin, IPropertiesPlugin, IGroupEnumerationPlugin
 
-from Products.PlonePAS.interfaces.group import IGroupIntrospection, \
-    IGroupManagement
+from Products.PlonePAS.interfaces.group import IGroupIntrospection
 from Products.PlonePAS.interfaces.plugins import IMutablePropertiesPlugin
 from Products.PlonePAS.plugins.group import PloneGroup
 
@@ -151,6 +150,5 @@ classImplements(
     IMutablePropertiesPlugin,
     IRolesPlugin,
     IGroupIntrospection,
-    IGroupManagement,
     )
 InitializeClass(PloneLDAPPluginBaseMixin)